put functionality implemented to edit item

+++ api/migrations/20240807001526_create_item_table.js
@@ -5,10 +5,10 @@
 exports.up = function(knex) {
   return knex.schema.createTable('item_table', table => {
     table.increments(); // adds an auto incrementing PK column
-    table.integer('userid').notNullable(); // equivalent of varchar(255)
+    table.bigInteger('userid').notNullable(); // equivalent of varchar(255)
     table.string('item_name').notNullable(); // equivalent of varchar(255)
     table.string('description', 1000).notNullable(); // equivalent of varchar(255)
-    table.integer('quantity').notNullable(); // equivalent of varchar(255)
+    table.bigInteger('quantity').notNullable(); // equivalent of varchar(255)
   });
 };
